[PRJ] Move UriLoader to XAG.Wpf
