+++ XamlAnimatedGif.Shared/Decoding/UnknownExtensionTypeException.cs
@@ -6,6 +6,5 @@ namespace XamlAnimatedGif.Decoding
     {
         internal UnknownExtensionTypeException(string message) : base(message) { }
         internal UnknownExtensionTypeException(string message, Exception inner) : base(message, inner) { }
- 
     }
 }
